+++ src/parser/mod.rs
@@ -9,4 +9,5 @@ pub(crate) mod parser;
 
 pub use errors::JsonPathParserError;
 pub use model::JsonPath;
+pub use model::JsonPathIndex;
 pub use parser::{parse_json_path, Rule};